// orbito-frontend/src/core/api/generated/models/updateClientCommand.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Orbito API
 * OpenAPI spec version: v1
 */

export interface UpdateClientCommand {
  id?: string;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  directEmail?: string | null;
  /** @nullable */
  directFirstName?: string | null;
  /** @nullable */
  directLastName?: string | null;
}